+++ src/script/main/globals.ts
@@ -36,8 +36,7 @@ import 'Components/Image';
 import 'Components/toggle/InfoToggle';
 import '../page/preferences/avPreferences/InputLevel';
 import 'Components/calling/FullscreenVideoCall';
-import 'Components/collection/CollectionDetails';
-import 'Components/collection/Collection';
+import '../page/collection/Collection';
 import 'Components/list/ConversationListCallingCell';
 import 'Components/list/ConversationListCell';
 import 'Components/avatar/GroupAvatar';
