+++ src/script/page/collection/CollectionDetails.test.tsx
