+++ src/script/page/collection/CollectionDetails.tsx
